+++ src/graphdatascience/session/remote_ops/projection_runner.py
@@ -44,10 +44,10 @@ class ProjectionRunner:
             )
             query = query.replace("gds.graph.project", "gds.graph.project.remote")
 
-
         if "undirectedRelationshipTypes" in query or "inverseIndexedRelationshipTypes" in query:
-            raise ValueError("The `undirectedRelationshipTypes` and `inverseIndexedRelationshipTypes` parameters need to be specified as separate arguments to `gds.graph.project.cypher` instead of inside the projection query.")
-
+            raise ValueError(
+                "The `undirectedRelationshipTypes` and `inverseIndexedRelationshipTypes` parameters need to be specified as separate arguments to `gds.graph.project.cypher` instead of inside the projection query."
+            )
 
         actual_job_id, query_runner = self._project_protocol.start_cypher_projection(
             graph_name,